Class hierarchy rework
Turned ARenderable into a class that inherits from GameObject.
GameObject will be the root class of all game object instances.

This necessitated the change of all instances and subclasses of
ARenderable to switch to receiving a GameEngine instance in their
constructor.

The GameObject will probably change the signature of Setup to avoid
relying on a GameModule (maybe).

AGL classes that don't inherit from ARenderable still receive an
instance of ARenderer (where applicable).

Removed the Draw() function of GameObject as this role
is given to inheritors of ARenderable. GameObjects should be
interactive but not necessarily drawable.

+++ engine/include/Asset.h
@@ -9,7 +9,7 @@
 #endif
 
 #define LARGE_FILE_MMAP_TRESH (1024*1024*10)// 10MB
-#define LARGE_FILE_MMAP_LIMIT (1024*1024*1024*4) // 4GB
+#define LARGE_FILE_MMAP_LIMIT ((uint64_t)1024*(uint64_t)1024*(uint64_t)1024*(uint64_t)4) // 4GB
 
 /**
  * \brief Type of an Asset.
